Added appropriate casts in 3 places to silence VS C warnings. These are an improvement for the human reader, too. 	modified:   src/lib/libdwarf/dwarf_crc32.c 	modified:   src/lib/libdwarf/dwarf_line_table_reader_common.h 	modified:   src/lib/libdwarf/dwarf_xu_index.c

+++ src/lib/libdwarf/dwarf_crc32.c
@@ -142,7 +142,8 @@ dwarf_crc32 (Dwarf_Debug dbg,unsigned char *crcbuf,
             return DW_DLV_ERROR;
         }
         /*  Call the public API function so it gets tested too. */
-        tcrc = (unsigned int)dwarf_basic_crc32(readbuf,readlenu,
+        tcrc = (unsigned int)dwarf_basic_crc32(readbuf,
+            (unsigned long)readlenu,
             (unsigned long)init);
         init = tcrc;
         size_left -= (off_t)readlenu;
